Add files via upload
For better accuracy of your model you shold keep different samples inside your training and test data_set. Change the dense of the last layer of the nn model according to your sample size. I had took 5 type of sample to train and feed the network that's why my dense of the last layer is 5.